chore: create folders for context in entity/database
