+++ internal/entity/user/user_test.go
@@ -1,4 +1,4 @@
-package entity
+package user
 
 import (
 	"testing"
